remove Guardian model
