+++ app/src/main/java/mhealth/mvax/records/record/patient/view/ViewPatientAdapter.java
@@ -43,8 +43,8 @@ public class ViewPatientAdapter extends PatientDetailsAdapter {
     // Constructors
     //================================================================================
 
-    ViewPatientAdapter(Context context, LinkedHashMap<Integer, List<Detail>> sectionedData) {
-        super(context, sectionedData);
+    ViewPatientAdapter(Context context, List<Detail> details) {
+        super(context, details);
     }
 
     //================================================================================
@@ -54,11 +54,11 @@ public class ViewPatientAdapter extends PatientDetailsAdapter {
     @Override
     public View getView(int position, View rowView, ViewGroup viewGroup) {
         int rowType = getItemViewType(position);
-        switch (rowType) {
-            case TYPE_SECTION: // section header
-                rowView = getSectionHeaderView(position, viewGroup);
-                break;
-            case TYPE_FIELD: // field displaying Detail
+//        switch (rowType) {
+//            case TYPE_SECTION: // section header
+//                rowView = getSectionHeaderView(position, viewGroup);
+//                break;
+//            case TYPE_FIELD: // field displaying Detail
                 rowView = mInflater.inflate(R.layout.list_item_record_detail, viewGroup, false);
 
                 final TextView fieldView = rowView.findViewById(R.id.textview_field);
@@ -67,8 +67,8 @@ public class ViewPatientAdapter extends PatientDetailsAdapter {
                 final EditText valueView = rowView.findViewById(R.id.textview_value);
                 valueView.setText(getItem(position).getStringValue()); // set field value
                 valueView.setFocusable(false); // EditText not editable in view record mode
-                break;
-        }
+//                break;
+//        }
         return rowView;
     }
 }
